+++ package.json
@@ -90,11 +90,11 @@
   },
   "devDependencies": {
     "@babel/core": "^7.3.3",
-    "@storybook/addon-actions": "^5.0.0-rc.3",
-    "@storybook/addon-links": "^5.0.0-rc.3",
-    "@storybook/addons": "^5.0.0-rc.3",
-    "@storybook/cli": "^5.0.0-rc.3",
-    "@storybook/react": "^5.0.0-rc.3",
+    "@storybook/addon-actions": "^5.0.0",
+    "@storybook/addon-links": "^5.0.0",
+    "@storybook/addons": "^5.0.0",
+    "@storybook/cli": "^5.0.0",
+    "@storybook/react": "^5.0.0",
     "@types/color": "^3.0.0",
     "@types/enzyme": "^3.1.18",
     "@types/enzyme-adapter-react-16": "^1.0.4",
